% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_spec.R
\name{chk_pth}
\alias{chk_pth}
\title{Check to see if provided spec is path or a JSON string.}
\usage{
chk_pth(file_or_json)
}
\arguments{
\item{file_or_json}{A string corresponding to either a filepath containing
the specification, or the specification itself as a json string}
}
\value{
A string identifying if the file was found or not
}
\description{
Check to see if provided spec is path or a JSON string.
}
