% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{coerce_dates}
\alias{coerce_dates}
\title{Helper function to coerce YYYYMMDD columns in a data frame to dates where
possible, using \code{lubridate::ymd()}. This function only converts columns
where every value successfully coerces to a date.}
\usage{
coerce_dates(table)
}
\arguments{
\item{table}{A data frame}
}
\value{
A data frame with columns coerced to dates where possible
}
\description{
Helper function to coerce YYYYMMDD columns in a data frame to dates where
possible, using \code{lubridate::ymd()}. This function only converts columns
where every value successfully coerces to a date.
}
