% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{context_message}
\alias{context_message}
\title{Concatenate context into string and append a message}
\usage{
context_message(context, message)
}
\arguments{
\item{context}{A character vector}

\item{message}{A single string}
}
\value{
result A single string
}
\description{
Concatenate context into string and append a message
}
