% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eider-package.R
\docType{package}
\name{eider-package}
\alias{eider}
\alias{eider-package}
\title{eider: Todo Acronym Has Not Been Decided}
\description{
ooooooOOOOooooo!!
}
\author{
\strong{Maintainer}: Camila Rangel Smith \email{crangelsmith@turing.ac.uk} [contributor]

Authors:
\itemize{
  \item Helen Duncan Little \email{hduncan@turing.ac.uk}
  \item Jonathan Yong \email{jyong@turing.ac.uk}
}

Other contributors:
\itemize{
  \item Catalina Vallejos \email{catalina.vallejos@ed.ac.uk} [contributor]
  \item Louis Aslett \email{louis.aslett@durham.ac.uk} [contributor]
  \item Simon Rogers \email{simon.rogers@nhs.scot} [contributor]
}

}
