% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{eider_example}
\alias{eider_example}
\title{Return path to the example datasets provided with the package}
\usage{
eider_example(file = NULL)
}
\arguments{
\item{file}{Name of the file to be used from inst/extdata.
If NULL all files will be listed}
}
\value{
A character string with the path to the file, or a vector of strings
}
\description{
Return path to the example datasets provided with the package
}
\examples{
eider_example()
eider_example("random_ae_data.csv")
}
