% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_context.R
\name{error_context}
\alias{error_context}
\title{Issue an error with a context and stop}
\usage{
error_context(context)
}
\arguments{
\item{context}{A character vector}

\item{err_message}{A string}
}
\description{
Issue an error with a context and stop
}
