% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featurise_present.R
\name{featurise_present}
\alias{featurise_present}
\title{Generates a feature which, for each patient ID, contains a 1 if any row that
passes a specified filter exists, or 0 if not.}
\usage{
featurise_present(all_tables, spec, context = NULL)
}
\arguments{
\item{all_tables}{List of all input tables (passed in from read_data).}

\item{spec}{A list containing the following elements:
\itemize{
\item source_file:         Filename of the source table to read from.
\item primary_filter:      A filter object to apply to the source table.
\item output_feature_name: Name of the output column.
\item grouping_columns:    Name of the column(s) to group by.
}}

\item{context}{A character vector to be used in logging or error messages.
Defaults to NULL.}
}
\value{
A list with the following elements:
\itemize{
\item feature_table: A data frame with a 1 or 0 in the output column for each
patient ID, depending on whether any rows in the source
table pass the filter.
\item missing_value: The value to use for patients who have no matching rows in
the source table. This value is passed downstream to the
function which joins all the feature tables together. By
definition, this value is 0.
}
}
\description{
Generates a feature which, for each patient ID, contains a 1 if any row that
passes a specified filter exists, or 0 if not.
}
