% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featurise_sum.R
\name{featurise_sum}
\alias{featurise_sum}
\title{Groups by ID and computes the sum of the values in all rows which pass a
given filter.}
\usage{
featurise_sum(all_tables, spec, context = NULL)
}
\arguments{
\item{all_tables}{List of all input tables (passed in from read_data).}

\item{spec}{A list containing the following elements:
\itemize{
\item source_file:         Filename of the source table to read from.
\item primary_filter:      A filter object to apply to the source table.
\item aggregation_column:  Name of the column which provides the values to be
\item 

\if{html}{\out{<div class="sourceCode">}}\preformatted{                 summed over.
}\if{html}{\out{</div>}}
\item output_feature_name: Name of the output column.
\item grouping_columns:    Name of the columns to group the source table by
before summation.
\item absent_default_value:The value to use for patients who have no matching
rows in the source table.
}}

\item{context}{A character vector to be used in logging or error messages.
Defaults to NULL.}
}
\value{
A list with the following elements:
\itemize{
\item feature_table: A data frame with one row per patient ID and one column
containing the sum of matching rows in the source table.
The column names are 'id' for the ID (this is standardised
across all feature tables), and the value of
output_column_name.
\item missing_value: The value to use for patients who have no matching rows in
the source table. This value is passed downstream to the
function which joins all the feature tables together.
}
}
\description{
Groups by ID and computes the sum of the values in all rows which pass a
given filter.
}
