% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_or.R
\name{filter_or}
\alias{filter_or}
\title{Filter table rows based on a logical disjunction of multiple filters (i.e.
filter1 OR filter2 OR ...)}
\usage{
filter_or(table, filter_obj, context = NULL)
}
\arguments{
\item{table}{A data frame}

\item{filter_obj}{A list containing the following elements:
- type: must be 'or' (case-insensitive)
- subfilters: a list of filter objects}

\item{context}{A string to be used in logging or error messages. Defaults to
NULL.}
}
\value{
A list with the following elements:
- passed: data frame with the rows that passed the filter
- rejected: all other rows
}
\description{
Filter table rows based on a logical disjunction of multiple filters (i.e.
filter1 OR filter2 OR ...)
}
