% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{log_context}
\alias{log_context}
\title{Log a message showing the current execution context}
\usage{
log_context(context, severity, message = "")
}
\arguments{
\item{context}{A character vector}

\item{severity}{A string in c("fatal", "error", "warn", "info", "debug",
"trace")}
}
\description{
Log a message showing the current execution context
}
