% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pad_missing_values.R
\name{pad_missing_values}
\alias{pad_missing_values}
\title{Given an ID column in a table, and a missing value, appends id-value pairs
for all values that are not in a second table. It is assumed that the IDs
are added to the 'id' column in the second table.}
\usage{
pad_missing_values(table, id_col, missing_value, id_table)
}
\arguments{
\item{table}{The table containing all the values that should be present}

\item{id_col}{The name of the column containing the IDs}

\item{missing_value}{The value that should be present for all IDs}

\item{id_table}{The table to be padded}
}
\value{
A table like \code{id_table}, but with rows appended for each ID that is
present in \code{table} but not present in \code{id_table}.
}
\description{
Given an ID column in a table, and a missing value, appends id-value pairs
for all values that are not in a second table. It is assumed that the IDs
are added to the 'id' column in the second table.
}
