% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_json.R
\name{parse_nested_filter}
\alias{parse_nested_filter}
\title{Parse a nested filter, returns a list of type (AND/OR) followed by lists
of the singular filters}
\usage{
parse_nested_filter(nested_filter)
}
\arguments{
\item{nested_filter}{A nested filter as defined in the origin json file}
}
\value{
A nested filter object
}
\description{
Parse a nested filter, returns a list of type (AND/OR) followed by lists
of the singular filters
}
