% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_json.R
\name{parse_single_or_nested}
\alias{parse_single_or_nested}
\title{Check if a filter is nested or single and parse accordingly}
\usage{
parse_single_or_nested(filter)
}
\description{
Check if a filter is nested or single and parse accordingly
}
