% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_table.R
\name{preprocess_table}
\alias{preprocess_table}
\title{Perform (date) preprocessing on a given data table}
\usage{
preprocess_table(input_table, spec, context = NULL)
}
\arguments{
\item{input_table}{The original input table which will be mutated}

\item{spec}{The specification which dictates the changes}

\item{context}{Execution context, defaulting to NULL}
}
\value{
A mutated version of the initial input table
}
\description{
Perform (date) preprocessing on a given data table
}
