% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_data}
\alias{read_data}
\title{Process the list of tables given, returning a list of tibbles. Each entry in
the list can either be a tibble itself (or a data.frame), or a string, in
which case it is assumed to be a file and is read in.}
\usage{
read_data(source_names)
}
\arguments{
\item{filenames}{A vector of strings or dataframe-like objects.}
}
\value{
A list of data frames. The names of the list are the same as the
names of the input list.
}
\description{
Process the list of tables given, returning a list of tibbles. Each entry in
the list can either be a tibble itself (or a data.frame), or a string, in
which case it is assumed to be a file and is read in.
}
