% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_one_table}
\alias{read_one_table}
\title{Helper function to read a single table from a file.}
\usage{
read_one_table(filepath_or_df, name)
}
\description{
Helper function to read a single table from a file.
}
