% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_spec.R
\name{read_spec_type}
\alias{read_spec_type}
\title{Read spec type and give appropriate logging messages}
\usage{
read_spec_type(filename_or_json_str)
}
\arguments{
\item{filename_or_json_str}{A string which may or may not be a valid filepath}
}
\value{
information if a file or a string has been provided
}
\description{
Read spec type and give appropriate logging messages
}
