% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{stop_context}
\alias{stop_context}
\title{On top of logging a FATAL message, stop execution entirely, and print a
message showing the current execution context}
\usage{
stop_context(context, message = "")
}
\description{
On top of logging a FATAL message, stop execution entirely, and print a
message showing the current execution context
}
