% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validate_absent_default_value}
\alias{validate_absent_default_value}
\title{Checks that \code{spec$absent_default_value} is a single number. If so, returns
the number.}
\usage{
validate_absent_default_value(spec, context)
}
\description{
Checks that \code{spec$absent_default_value} is a single number. If so, returns
the number.
}
