% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validate_column_present}
\alias{validate_column_present}
\title{Checks that \code{spec[[field_name]]} is a single string and that it is a column
in the table \code{tbl}. If so, returns the column name.}
\usage{
validate_column_present(field_name, spec, tbl, context)
}
\description{
Checks that \code{spec[[field_name]]} is a single string and that it is a column
in the table \code{tbl}. If so, returns the column name.
}
