% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validate_columns_present}
\alias{validate_columns_present}
\title{Same as validate_column_present, but allows for multiple columns to be
specified.}
\usage{
validate_columns_present(field_name, spec, tbl, context)
}
\description{
Same as validate_column_present, but allows for multiple columns to be
specified.
}
