% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validate_filter_column}
\alias{validate_filter_column}
\title{Checks that the column specified by a filter object exists in a table. If
so, returns the column name.}
\usage{
validate_filter_column(filter_obj, tbl, context)
}
\description{
Checks that the column specified by a filter object exists in a table. If
so, returns the column name.
}
