% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validate_filter_value}
\alias{validate_filter_value}
\title{Checks that the values specified by a filter object are of the same type as
the column to be filtered on. If so, returns the values.}
\usage{
validate_filter_value(filter_obj, column, context)
}
\description{
Checks that the values specified by a filter object are of the same type as
the column to be filtered on. If so, returns the values.
}
