% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validate_output_feature_name}
\alias{validate_output_feature_name}
\title{Checks that \code{spec$output_feature_name} is a single string. If so, returns
the string.}
\usage{
validate_output_feature_name(spec, context)
}
\description{
Checks that \code{spec$output_feature_name} is a single string. If so, returns
the string.
}
