% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validate_source_file}
\alias{validate_source_file}
\title{Checks that \code{spec$source_file} is a single string and that it is a key in
\code{all_tables}. If so, returns the table itself (or the file path of the
table).}
\usage{
validate_source_file(spec, all_tables, context)
}
\description{
Checks that \code{spec$source_file} is a single string and that it is a key in
\code{all_tables}. If so, returns the table itself (or the file path of the
table).
}
