% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eider-package.R
\docType{package}
\name{eider-package}
\alias{eider}
\alias{eider-package}
\title{eider: Declarative Feature Extraction from Tabular Data Records}
\description{
\code{eider} is a lightweight package for processing tabular data in a
declarative fashion. To get started, see:
\code{vignette("eider")}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/alan-turing-institute/eider}
  \item Report bugs at \url{https://github.com/alan-turing-institute/eider/issues}
}

}
\author{
\strong{Maintainer}: Camila Rangel Smith \email{crangelsmith@turing.ac.uk} (\href{https://orcid.org/0000-0002-0227-836X}{ORCID}) [contributor]

Authors:
\itemize{
  \item Helen Duncan Little \email{hduncan@turing.ac.uk} (\href{https://orcid.org/0000-0002-0897-7188}{ORCID})
  \item Jonathan Yong \email{jyong@turing.ac.uk} (\href{https://orcid.org/0000-0002-2472-974X}{ORCID})
}

Other contributors:
\itemize{
  \item Catalina Vallejos \email{catalina.vallejos@ed.ac.uk} (\href{https://orcid.org/0000-0003-3638-1960}{ORCID}) [contributor]
  \item Louis Aslett \email{louis.aslett@durham.ac.uk} (\href{https://orcid.org/0000-0003-2211-233X}{ORCID}) [contributor]
  \item Simon Rogers \email{simon.rogers@nhs.scot} (\href{https://orcid.org/0000-0003-3578-4477}{ORCID}) [contributor]
  \item The Alan Turing Institute [copyright holder, funder]
}

}
\keyword{internal}
