% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{eider_example}
\alias{eider_example}
\title{Obtain filepaths to example data and JSON features}
\usage{
eider_example(file = NULL)
}
\arguments{
\item{file}{The filename to return the full path for. Defaults to \code{NULL}, in
which case it will return a vector of all valid filenames.}
}
\value{
A string containing the full path to the file, or a vector of
filenames
}
\description{
Return an absolute path to the example data and JSON features provided with
the package. These files are contained in the package \code{inst/extdata}
directory.
}
\examples{
eider_example()
eider_example("random_ae_data.csv")
}
