% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_pipeline.R
\name{run_pipeline}
\alias{run_pipeline}
\title{Perform the entire feature transformation process.}
\usage{
run_pipeline(
  data_sources,
  feature_filenames = NULL,
  response_filenames = NULL,
  all_ids = NULL
)
}
\arguments{
\item{data_sources}{A list mapping unique table identifiers to either
the file path from which they should be read from, or the data frame itself.}

\item{feature_filenames}{A vector of file paths to the feature JSON
specifications. Defaults to NULL.}

\item{response_filenames}{A vector of file paths to the feature JSON
specifications. Defaults to NULL.}

\item{all_ids}{A vector of all the unique identifiers that should be in the
final feature table. If not given, will generate a feature table containing
all unique identifiers found in input tables used by at least one feature.}
}
\value{
A data frame with the feature tables joined together
}
\description{
Perform the entire feature transformation process.
}
